# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the zODC Masterflow Service. It provides endpoints for managing workflows, requests, forms, and nodes.",
        "title": "zODC Masterflow Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1  // Adjusted BasePath for clarity (e.g., /api/v1) - Ensure routes match this prefix",
    "paths": {
...
        "/services/utils/excel-extract": {
            "post": {
                "description": "Uploads an Excel file and extracts data from it.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utils"
                ],
                "summary": "Extract data from an Excel file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: No file is received",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Fail extract excel file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/services/workflows": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of workflows based on various filter criteria for the logged-in user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Find all workflows",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by category ID",
                        "name": "categoryId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for workflows",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by workflow type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by project key",
                        "name": "projectKey",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by whether it has sub-workflows",
                        "name": "hasSubWorkflow",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by archival status",
                        "name": "isArchived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.WorkflowResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Error message for internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds a new workflow definition based on the provided data. Requires user authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Create a new workflow",
                "parameters": [
                    {
                        "description": "Workflow Creation Request",
                        "name": "workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateWorkflow"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: Workflow created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Error message for bad request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error message for internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/services/workflows/start": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Initiates a new instance of a workflow based on the provided request details. Requires user authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Start a workflow instance",
                "parameters": [
                    {
                        "description": "Start Workflow Request",
                        "name": "startRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.StartWorkflow"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error message for bad request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error message for internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/services/workflows/{id}": {
            "get": {
                "description": "Retrieves detailed information for a specific workflow version ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Find workflow detail by version ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow Version ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.WorkflowDetailResponse"
                        }
                    },
                    "500": {
                        "description": "Error message for invalid ID or internal server error\" // Combined 400/500 based on current code",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates a workflow definition based on the provided request details. Requires user authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Update a workflow",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Workflow Request",
                        "name": "workflow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateWorkflow"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Workflow updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error message for invalid ID or internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/services/workflows/{id}/archive": {
            "patch": {
                "description": "Marks a workflow definition as archived based on its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Archive a workflow",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Workflow archived successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error message for invalid ID or internal server error\" // Combined 400/500 based on current code",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "requests.CreateWorkflow": {
            "type": "object"
        },
...
        "requests.StartWorkflow": {
            "type": "object"
        },
...
        "requests.UpdateWorkflow": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "decoration": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "projectKey": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responses.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responses.ConnectionResponse": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
        "responses.NodeDataConditionResponse": {
            "type": "object",
            "properties": {
                "falseDestinations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trueDestinations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "responses.NodeDataResponse": {
            "type": "object",
            "properties": {
                "actualEndTime": {
                    "type": "string"
                },
                "actualStartTime": {
                    "type": "string"
                },
                "assignee": {
                    "$ref": "#/definitions/types.Assignee"
                },
                "attachFiles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "condition": {
                    "$ref": "#/definitions/responses.NodeDataConditionResponse"
                },
                "description": {
                    "type": "string"
                },
                "editorContent": {
                    "$ref": "#/definitions/responses.NodeDataResponseEditorContent"
                },
                "endDate": {
                    "type": "string"
                },
                "endType": {
                    "type": "string"
                },
                "estimatePoint": {
                    "type": "number"
                },
                "formAttached": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.NodeFormResponse"
                    }
                },
                "jiraLinkUrl": {
                    "type": "string"
                },
                "plannedEndTime": {
                    "type": "string"
                },
                "plannedStartTime": {
                    "description": "Add for update request",
                    "type": "string"
                },
                "subRequestId": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responses.NodeFormDataResponse": {
            "type": "object",
            "properties": {
                "fieldId": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "responses.NodeResponse": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/responses.NodeDataResponse"
                },
                "form": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.NodeFormDataResponse"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isCurrent": {
                    "type": "boolean"
                },
                "jiraKey": {
                    "type": "string"
                },
                "lastSyncedAt": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "parentId": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/types.Position"
                },
                "size": {
                    "$ref": "#/definitions/types.Size"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responses.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.StoryResponse": {
            "type": "object",
            "properties": {
                "categoryKey": {
                    "type": "string"
                },
                "decoration": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "isSystemLinked": {
                    "type": "boolean"
                },
                "node": {
                    "$ref": "#/definitions/responses.NodeResponse"
                },
                "progress": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responses.WorkflowDetailResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/responses.CategoryResponse"
                },
                "connections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ConnectionResponse"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "currentVersion": {
                    "type": "integer"
                },
                "decoration": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isArchived": {
                    "type": "boolean"
                },
                "isTemplate": {
                    "type": "boolean"
                },
                "lastAssignee": {
                    "$ref": "#/definitions/types.Assignee"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.NodeResponse"
                    }
                },
                "projectKey": {
                    "type": "string"
                },
                "requestId": {
                    "type": "integer"
                },
                "stories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.StoryResponse"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responses.WorkflowResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/responses.CategoryResponse"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentVersion": {
                    "type": "integer"
                },
                "decoration": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isArchived": {
                    "type": "boolean"
                },
                "projectKey": {
                    "type": "string"
                },
                "requestId": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "types.Assignee": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isSystemUser": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.Position": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "types.Size": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "number"
                },
                "width": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}